refactor(gui): swap graph layout engine elkjs→@dagrejs/dagre
elkjs 是 EPL-2.0（弱 copyleft，与 AGPL 不兼容）→ 换 MIT 的 @dagrejs/dagre。
graphLayout.ts 用 dagre compound graph（module=cluster，entity=child，dagre 计算节点位置+cluster 包围盒），React Flow 节点/边渲染、颜色、边样式、cycle 检测全部不变。GraphView 只改 import 与文案（Powered by dagre）。实测真台账 283 节点经 dagre 渲染、0 console error。
